Add favicons and images for certain shared links/pages.
