+++ app/layout.tsx
@@ -5,6 +5,14 @@ import type { ReactNode } from 'react';
 export const metadata: Metadata = {
   title: 'Stroman Properties',
   description: 'Luxury vacation rentals',
+  icons: {
+    icon: [
+      { url: '/images/favicon-16x16.png', sizes: '16x16', type: 'image/png' },
+      { url: '/images/favicon-32x32.png', sizes: '32x32', type: 'image/png' },
+      { url: '/images/favicon-48x48.png', sizes: '48x48', type: 'image/png' },
+      { url: '/images/favicon-64x64.png', sizes: '64x64', type: 'image/png' },
+    ],
+  },
 };
 
 export const viewport: Viewport = {
